BAsicChatbot/requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
chromadb
langchain-text-splitters
beautifulsoup4
requests
pypdf
python-dotenv
python-docx
langsmith
